Use unsafe to pass World as reference rather than Arc<World>
This is necessary for the command buffers to be applied between ticks,
since they need mutable access
to `World` and `Arc` doesn't allow that.

Alternatives that were thoroughly investigated
before falling back on `unsafe`:
* Using Arc<RwLock<World>>. This is extremely difficult due to
all the places in the engine which require world-references
* Using lock-free datastructures inside `World` to make it not require
`&mut self` (interior mutability)
* Making `World` contain an `RwLock<InnerWorld>` which then contains all
the current `World`-data.
This is not viable since functions like `borrow_component_vecs` need to
return references to data within `InnerWorld` and that leaks the
RwLockGuards, leading to the same problem as alternative #1.